# First iteration on performance issues on edit registrations page