See #899.

With ~200 registrations (using modified development seeds) for a competition I get the following numbers for http://localhost:3000/competitions/MyComp5002018/edit/registrations using the production environment (timed using network tab in Chrome):
```
master: 9.7s, 10.2s, 9.6s, 9.7s
this branch: 585ms, 516ms, 522ms, 517ms
```

There are still a bunch of SQL calls to the `countries` table which we should remove as well, but at least this PR should fix the production errors.

What I did: Create a matrix of the form `{0,1}^(|events| x |registrations|)` before rendering the form. `1` means the competitor has registered for that event, `0` means he didn't. This representation had the advantage that I could reuse it for the footer (where we count the number of people who've registered for each event) by simply transposing the matrix and summing over each row.