@timhabermaas 2016-09-17
Closing in favor of #904.

I've checked @viroulep's branch and his performance gains are pretty much the same as mine (also ~500ms per request for the same data). My solution is less idomatic, though.